requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2023.10.0
future==0.18.3
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper @ git+https://github.com/openai/whisper.git@8bc8860694949db53c42ba47ddc23786c2e02a8b
Pillow==10.1.0
proglog==0.1.10
regex==2023.10.3
requests==2.31.0
sympy==1.12
tiktoken==0.5.1
torch==2.1.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
